+++ Samples/Playlists/cs/FilterViewModel/FilterProductCriteria.cs
@@ -8,11 +8,11 @@ namespace SDKTemplate
 {
     public class FilterProductCriteria
     {
-        public IRange<float> DiscountPerRange { get; set; }
+        public IRange<decimal> DiscountPerRange { get; set; }
         public IRange<Int32> QuantityRange { get; set; }
         public bool IncludeDeficientItemsOnly { get; set; }
 
-        public FilterProductCriteria(IRange<float> discounPerRange, IRange<Int32> quantityRange, bool? isChecked)
+        public FilterProductCriteria(IRange<decimal> discounPerRange, IRange<Int32> quantityRange, bool? isChecked)
         {
             this.DiscountPerRange = discounPerRange;
             this.QuantityRange = quantityRange;
